3.6.0 b640
work in progress

+++ core/src/main/java/me/lokka30/levelledmobs/LevelledMobs.java
@@ -72,7 +72,7 @@ public final class LevelledMobs extends JavaPlugin {
     public PlaceholderApiIntegration placeholderApiIntegration;
     public boolean migratedFromPre30;
     public YmlParsingHelper helperSettings;
-    public int playerLevellingMinRelevelTime;
+    public long playerLevellingMinRelevelTime;
     public int maxPlayersRecorded;
 
     // Configuration
